requirements.txt
numpy==2.2.1
torch==2.5.1
pandas==2.2.3
seaborn==0.13.2
matplotlib==3.10.0
xlrd==2.0.1 # for reading excel files
scipy==1.15.1 
xgboost==2.1.4
scikit-learn==1.6.1
imblearn==0.0 # for handling imbalanced data
graphviz==0.20.3 # for plotting the individual trees of the xgboost model
scikit-multilearn==0.2.0
sklearn-hierarchical-classification==1.3.2
modal==0.73.53 # for active learning